feat(pack): sync latest next.js
