+++ crates/pack-api/src/project.rs
@@ -615,11 +615,11 @@ impl Issue for ConflictIssue {
     }
 
     async fn title(&self) -> Result<StyledString> {
-        Ok((*self.title.await?).clone())
+        self.title.owned().await
     }
 
     async fn description(&self) -> Result<Option<StyledString>> {
-        Ok(Some((*self.description.await?).clone()))
+        Ok(Some(self.description.owned().await?))
     }
 }
 
@@ -1101,11 +1101,14 @@ impl Project {
                 .copied()
                 .map(ResolvedVc::upcast)
                 .collect();
-            ModuleGraph::from_single_graph(SingleModuleGraph::new_with_entries(
-                ResolvedVc::cell(vec![ChunkGroupEntry::Entry(entries)]),
-                is_production,
-                is_production,
-            ))
+            ModuleGraph::from_graphs(
+                vec![SingleModuleGraph::new_with_entries(
+                    ResolvedVc::cell(vec![ChunkGroupEntry::Entry(entries)]),
+                    is_production,
+                    is_production,
+                )],
+                None,
+            )
             .connect()
         } else {
             *self.whole_app_module_graphs().await?.full
@@ -1300,14 +1303,15 @@ impl Project {
             .copied()
             .map(ResolvedVc::upcast)
             .collect();
-        Ok(
-            ModuleGraph::from_single_graph(SingleModuleGraph::new_with_entries(
+        Ok(ModuleGraph::from_graphs(
+            vec![SingleModuleGraph::new_with_entries(
                 ResolvedVc::cell(vec![ChunkGroupEntry::Entry(entries)]),
                 is_production,
                 is_production,
-            ))
-            .connect(),
+            )],
+            None,
         )
+        .connect())
     }
 
     #[turbo_tasks::function]
@@ -1714,7 +1718,7 @@ async fn whole_app_module_graph_operation(
     );
     let base_visited_modules = VisitedModules::from_graph(base_single_module_graph);
 
-    let base = ModuleGraph::from_single_graph(base_single_module_graph);
+    let base = ModuleGraph::from_graphs(vec![base_single_module_graph], None);
 
     let remove_unused_imports = *project.config().remove_unused_imports(mode).await?;
 
@@ -1722,10 +1726,7 @@ async fn whole_app_module_graph_operation(
         // TODO suboptimal that we do compute_binding_usage_info twice (once for the base graph
         // and later for the full graph)
         let binding_usage_info = compute_binding_usage_info(base, true);
-        ModuleGraph::from_single_graph_without_unused_references(
-            base_single_module_graph,
-            binding_usage_info,
-        )
+        ModuleGraph::from_graphs(vec![base_single_module_graph], Some(binding_usage_info))
     } else {
         base
     };
@@ -1745,14 +1746,14 @@ async fn whole_app_module_graph_operation(
     let graphs = vec![base_single_module_graph, additional_module_graph];
 
     let (full, binding_usage_info) = if remove_unused_imports {
-        let full_with_unused_references = ModuleGraph::from_graphs(graphs.clone());
+        let full_with_unused_references = ModuleGraph::from_graphs(graphs.clone(), None);
         let binding_usage_info = compute_binding_usage_info(full_with_unused_references, true);
         (
-            ModuleGraph::from_graphs_without_unused_references(graphs, binding_usage_info),
+            ModuleGraph::from_graphs(graphs, Some(binding_usage_info)),
             Some(binding_usage_info),
         )
     } else {
-        (ModuleGraph::from_graphs(graphs), None)
+        (ModuleGraph::from_graphs(graphs, None), None)
     };
 
     Ok(BaseAndFullModuleGraph {
