+++ crates/pack-tests/tests/snapshot/externals/async-script-externals/output/_root-of-the-server___d0ad72cc.js
@@ -2,7 +2,8 @@
 "[externals]/JSZip [external] (JSZip@https://gw.alipayobjects.com/os/lib/jszip/3.10.1/dist/jszip.min.js, script)", ((__turbopack_context__) => {
 "use strict";
 
-return __turbopack_context__.a(async (__turbopack_handle_async_dependencies__, __turbopack_async_result__) => { try {
+return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
+try {
 var mod = await (async () => {
   if (typeof globalThis["JSZip"] !== 'undefined') {
     return globalThis["JSZip"];
@@ -23,7 +24,9 @@ for (var key in mod) ns[key] = mod[key];
 ns.default = mod;
 __turbopack_context__.n(ns);
 __turbopack_async_result__();
-} catch(e) { __turbopack_async_result__(e); } }, true);}),
+} catch(e) { __turbopack_async_result__(e); }
+}, true);
+}),
 "[externals]/_ [external] (_@https://gw.alipayobjects.com/os/lib/lodash/4.17.21/lodash.min.js, script, async loader)", ((__turbopack_context__) => {
 
 __turbopack_context__.v((parentImport) => {
@@ -37,26 +40,31 @@ __turbopack_context__.v((parentImport) => {
 "[project]/externals/async-script-externals/input/index.ts [client] (ecmascript)", ((__turbopack_context__) => {
 "use strict";
 
-return __turbopack_context__.a(async (__turbopack_handle_async_dependencies__, __turbopack_async_result__) => { try {
-var __TURBOPACK__imported__module__$5b$externals$5d2f$JSZip__$5b$external$5d$__$28$JSZip$40$https$3a2f2f$gw$2e$alipayobjects$2e$com$2f$os$2f$lib$2f$jszip$2f$3$2e$10$2e$1$2f$dist$2f$jszip$2e$min$2e$js$2c$__script$29$__ = __turbopack_context__.i("[externals]/JSZip [external] (JSZip@https://gw.alipayobjects.com/os/lib/jszip/3.10.1/dist/jszip.min.js, script)");
-var __turbopack_async_dependencies__ = __turbopack_handle_async_dependencies__([
-    __TURBOPACK__imported__module__$5b$externals$5d2f$JSZip__$5b$external$5d$__$28$JSZip$40$https$3a2f2f$gw$2e$alipayobjects$2e$com$2f$os$2f$lib$2f$jszip$2f$3$2e$10$2e$1$2f$dist$2f$jszip$2e$min$2e$js$2c$__script$29$__
-]);
-[__TURBOPACK__imported__module__$5b$externals$5d2f$JSZip__$5b$external$5d$__$28$JSZip$40$https$3a2f2f$gw$2e$alipayobjects$2e$com$2f$os$2f$lib$2f$jszip$2f$3$2e$10$2e$1$2f$dist$2f$jszip$2e$min$2e$js$2c$__script$29$__] = __turbopack_async_dependencies__.then ? (await __turbopack_async_dependencies__)() : __turbopack_async_dependencies__;
-;
-const zip = new __TURBOPACK__imported__module__$5b$externals$5d2f$JSZip__$5b$external$5d$__$28$JSZip$40$https$3a2f2f$gw$2e$alipayobjects$2e$com$2f$os$2f$lib$2f$jszip$2f$3$2e$10$2e$1$2f$dist$2f$jszip$2e$min$2e$js$2c$__script$29$__["default"]();
-zip;
-const func = async ()=>{
-    // @ts-ignore
-    const _ = await __turbopack_context__.A("[externals]/_ [external] (_@https://gw.alipayobjects.com/os/lib/lodash/4.17.21/lodash.min.js, script, async loader)");
-    console.log(Object.keys(_.default.omit({
-        a: 1
-    }, 'a')).length === 0);
-};
-func();
-__turbopack_context__.s([]);
-__turbopack_async_result__();
-} catch(e) { __turbopack_async_result__(e); } }, false);}),
+return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
+    try {
+        var __TURBOPACK__imported__module__$5b$externals$5d2f$JSZip__$5b$external$5d$__$28$JSZip$40$https$3a2f2f$gw$2e$alipayobjects$2e$com$2f$os$2f$lib$2f$jszip$2f$3$2e$10$2e$1$2f$dist$2f$jszip$2e$min$2e$js$2c$__script$29$__ = __turbopack_context__.i("[externals]/JSZip [external] (JSZip@https://gw.alipayobjects.com/os/lib/jszip/3.10.1/dist/jszip.min.js, script)");
+        var __turbopack_async_dependencies__ = __turbopack_handle_async_dependencies__([
+            __TURBOPACK__imported__module__$5b$externals$5d2f$JSZip__$5b$external$5d$__$28$JSZip$40$https$3a2f2f$gw$2e$alipayobjects$2e$com$2f$os$2f$lib$2f$jszip$2f$3$2e$10$2e$1$2f$dist$2f$jszip$2e$min$2e$js$2c$__script$29$__
+        ]);
+        [__TURBOPACK__imported__module__$5b$externals$5d2f$JSZip__$5b$external$5d$__$28$JSZip$40$https$3a2f2f$gw$2e$alipayobjects$2e$com$2f$os$2f$lib$2f$jszip$2f$3$2e$10$2e$1$2f$dist$2f$jszip$2e$min$2e$js$2c$__script$29$__] = __turbopack_async_dependencies__.then ? (await __turbopack_async_dependencies__)() : __turbopack_async_dependencies__;
+        ;
+        const zip = new __TURBOPACK__imported__module__$5b$externals$5d2f$JSZip__$5b$external$5d$__$28$JSZip$40$https$3a2f2f$gw$2e$alipayobjects$2e$com$2f$os$2f$lib$2f$jszip$2f$3$2e$10$2e$1$2f$dist$2f$jszip$2e$min$2e$js$2c$__script$29$__["default"]();
+        zip;
+        const func = async ()=>{
+            // @ts-ignore
+            const _ = await __turbopack_context__.A("[externals]/_ [external] (_@https://gw.alipayobjects.com/os/lib/lodash/4.17.21/lodash.min.js, script, async loader)");
+            console.log(Object.keys(_.default.omit({
+                a: 1
+            }, 'a')).length === 0);
+        };
+        func();
+        __turbopack_context__.s([]);
+        __turbopack_async_result__();
+    } catch (e) {
+        __turbopack_async_result__(e);
+    }
+}, false);
+}),
 ]);
 
 //# sourceMappingURL=_root-of-the-server___d0ad72cc.js.map
