+++ crates/pack-tests/tests/snapshot/library/async-module/output/main.js
@@ -6,79 +6,104 @@
 263, ((__turbopack_context__) => {
 "use strict";
 
-return __turbopack_context__.a(async (__turbopack_handle_async_dependencies__, __turbopack_async_result__) => { try {
-await 1;
-await 1;
-var __TURBOPACK__default__export__ = "hello";
-__turbopack_context__.s([
-    "default",
-    ()=>__TURBOPACK__default__export__
-]);
-__turbopack_async_result__();
-} catch(e) { __turbopack_async_result__(e); } }, true);}),
+return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
+    try {
+        await 1;
+        await 1;
+        var __TURBOPACK__default__export__ = "hello";
+        __turbopack_context__.s([
+            "default",
+            ()=>__TURBOPACK__default__export__
+        ]);
+        __turbopack_async_result__();
+    } catch (e) {
+        __turbopack_async_result__(e);
+    }
+}, true);
+}),
 356, ((__turbopack_context__) => {
 "use strict";
 
-return __turbopack_context__.a(async (__turbopack_handle_async_dependencies__, __turbopack_async_result__) => { try {
-var __TURBOPACK__imported__module__263__ = __turbopack_context__.i(263);
-var __turbopack_async_dependencies__ = __turbopack_handle_async_dependencies__([
-    __TURBOPACK__imported__module__263__
-]);
-[__TURBOPACK__imported__module__263__] = __turbopack_async_dependencies__.then ? (await __turbopack_async_dependencies__)() : __turbopack_async_dependencies__;
-;
-var __TURBOPACK__default__export__ = __TURBOPACK__imported__module__263__["default"] + " world";
-__turbopack_context__.s([
-    "default",
-    ()=>__TURBOPACK__default__export__
-]);
-__turbopack_async_result__();
-} catch(e) { __turbopack_async_result__(e); } }, false);}),
+return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
+    try {
+        var __TURBOPACK__imported__module__263__ = __turbopack_context__.i(263);
+        var __turbopack_async_dependencies__ = __turbopack_handle_async_dependencies__([
+            __TURBOPACK__imported__module__263__
+        ]);
+        [__TURBOPACK__imported__module__263__] = __turbopack_async_dependencies__.then ? (await __turbopack_async_dependencies__)() : __turbopack_async_dependencies__;
+        ;
+        var __TURBOPACK__default__export__ = __TURBOPACK__imported__module__263__["default"] + " world";
+        __turbopack_context__.s([
+            "default",
+            ()=>__TURBOPACK__default__export__
+        ]);
+        __turbopack_async_result__();
+    } catch (e) {
+        __turbopack_async_result__(e);
+    }
+}, false);
+}),
 220, ((__turbopack_context__) => {
 "use strict";
 
-return __turbopack_context__.a(async (__turbopack_handle_async_dependencies__, __turbopack_async_result__) => { try {
-var __TURBOPACK__imported__module__513__ = __turbopack_context__.i(513);
-var __TURBOPACK__imported__module__356__ = __turbopack_context__.i(356);
-var __turbopack_async_dependencies__ = __turbopack_handle_async_dependencies__([
-    __TURBOPACK__imported__module__513__,
-    __TURBOPACK__imported__module__356__
-]);
-[__TURBOPACK__imported__module__513__, __TURBOPACK__imported__module__356__] = __turbopack_async_dependencies__.then ? (await __turbopack_async_dependencies__)() : __turbopack_async_dependencies__;
-;
-;
-var __TURBOPACK__default__export__ = __TURBOPACK__imported__module__513__["default"] + ", " + __TURBOPACK__imported__module__356__["default"];
-__turbopack_context__.s([
-    "default",
-    ()=>__TURBOPACK__default__export__
-]);
-__turbopack_async_result__();
-} catch(e) { __turbopack_async_result__(e); } }, false);}),
+return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
+    try {
+        var __TURBOPACK__imported__module__513__ = __turbopack_context__.i(513);
+        var __TURBOPACK__imported__module__356__ = __turbopack_context__.i(356);
+        var __turbopack_async_dependencies__ = __turbopack_handle_async_dependencies__([
+            __TURBOPACK__imported__module__513__,
+            __TURBOPACK__imported__module__356__
+        ]);
+        [__TURBOPACK__imported__module__513__, __TURBOPACK__imported__module__356__] = __turbopack_async_dependencies__.then ? (await __turbopack_async_dependencies__)() : __turbopack_async_dependencies__;
+        ;
+        ;
+        var __TURBOPACK__default__export__ = __TURBOPACK__imported__module__513__["default"] + ", " + __TURBOPACK__imported__module__356__["default"];
+        __turbopack_context__.s([
+            "default",
+            ()=>__TURBOPACK__default__export__
+        ]);
+        __turbopack_async_result__();
+    } catch (e) {
+        __turbopack_async_result__(e);
+    }
+}, false);
+}),
 511, ((__turbopack_context__) => {
 "use strict";
 
-return __turbopack_context__.a(async (__turbopack_handle_async_dependencies__, __turbopack_async_result__) => { try {
-const result = await __turbopack_context__.r(220);
-await Promise.resolve().then(()=>__turbopack_context__.i(864));
-result.default;
-__turbopack_async_result__();
-} catch(e) { __turbopack_async_result__(e); } }, true);}),
+return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
+    try {
+        const result = await __turbopack_context__.r(220);
+        await Promise.resolve().then(()=>__turbopack_context__.i(864));
+        result.default;
+        __turbopack_async_result__();
+    } catch (e) {
+        __turbopack_async_result__(e);
+    }
+}, true);
+}),
 513, ((__turbopack_context__) => {
 "use strict";
 
-return __turbopack_context__.a(async (__turbopack_handle_async_dependencies__, __turbopack_async_result__) => { try {
-var __TURBOPACK__imported__module__263__ = __turbopack_context__.i(263);
-var __turbopack_async_dependencies__ = __turbopack_handle_async_dependencies__([
-    __TURBOPACK__imported__module__263__
-]);
-[__TURBOPACK__imported__module__263__] = __turbopack_async_dependencies__.then ? (await __turbopack_async_dependencies__)() : __turbopack_async_dependencies__;
-;
-var __TURBOPACK__default__export__ = __TURBOPACK__imported__module__263__["default"] + " world";
-__turbopack_context__.s([
-    "default",
-    ()=>__TURBOPACK__default__export__
-]);
-__turbopack_async_result__();
-} catch(e) { __turbopack_async_result__(e); } }, false);}),
+return __turbopack_context__.a(async function(__turbopack_handle_async_dependencies__, __turbopack_async_result__) {
+    try {
+        var __TURBOPACK__imported__module__263__ = __turbopack_context__.i(263);
+        var __turbopack_async_dependencies__ = __turbopack_handle_async_dependencies__([
+            __TURBOPACK__imported__module__263__
+        ]);
+        [__TURBOPACK__imported__module__263__] = __turbopack_async_dependencies__.then ? (await __turbopack_async_dependencies__)() : __turbopack_async_dependencies__;
+        ;
+        var __TURBOPACK__default__export__ = __TURBOPACK__imported__module__263__["default"] + " world";
+        __turbopack_context__.s([
+            "default",
+            ()=>__TURBOPACK__default__export__
+        ]);
+        __turbopack_async_result__();
+    } catch (e) {
+        __turbopack_async_result__(e);
+    }
+}, false);
+}),
 864, ((__turbopack_context__) => {
 "use strict";
