+++ crates/pack-tests/tests/snapshot/target_web/legacy_mobile/input/async_module.js
@@ -1,0 +1,4 @@
+const data = await fetch("/api/config").then((r) => r.json());
+
+export const config = data;
+export const version = "1.0.0";
