+++ crates/pack-tests/tests/snapshot/target_web/legacy_mobile/output/main.js
@@ -1,5 +1,5 @@
 (globalThis["TURBOPACK"] || (globalThis["TURBOPACK"] = [])).push([
     typeof document === "object" ? document.currentScript : undefined,
-    {"otherChunks":["_project___5c265bc3.js"],"runtimeModuleIds":["[project]/target_web/legacy_mobile/input/index.js [client] (ecmascript)"]}
+    {"otherChunks":["_project___fdc3448b.js"],"runtimeModuleIds":["[project]/target_web/legacy_mobile/input/index.js [client] (ecmascript)"]}
 ]);
 // Dummy runtime
