+++ examples/utooweb-demo/package.json
@@ -9,8 +9,8 @@
     "@monaco-editor/react": "^4.4.6",
     "@utoo/web": "*",
     "jszip": "^3.10.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react": "^19.2.5",
+    "react-dom": "^19.2.5"
   },
   "devDependencies": {
     "css-loader": "^7.1.2",
