+++ packages/utoo-web/src/utoo/index.d.ts
@@ -208,8 +208,27 @@ export function workerCreated(worker_id: number): void;
 export type InitInput = RequestInfo | URL | Response | BufferSource | WebAssembly.Module;
 
 export interface InitOutput {
+    readonly getWasmMemory: () => any;
+    readonly getWasmModule: () => any;
     readonly initLogFilter: (a: number, b: number) => void;
     readonly init_pack: () => void;
+    readonly __wbg_buildoptions_free: (a: number, b: number) => void;
+    readonly __wbg_get_buildoptions_cleanup: (a: number) => number;
+    readonly __wbg_roottask_free: (a: number, b: number) => void;
+    readonly __wbg_set_buildoptions_cleanup: (a: number, b: number) => void;
+    readonly buildoptions_config: (a: number) => any;
+    readonly buildoptions_new: () => number;
+    readonly buildoptions_set_config: (a: number, b: any) => void;
+    readonly registerWorkerScheduler: (a: any, b: any) => void;
+    readonly workerCreated: (a: number) => void;
+    readonly ERR_ABORT: () => [number, number];
+    readonly ERR_INVALID_STATE: () => [number, number];
+    readonly ERR_KEY_ALREADY_EXISTS: () => [number, number];
+    readonly ERR_NOT_ALLOWED: () => [number, number];
+    readonly ERR_NOT_FOUND: () => [number, number];
+    readonly ERR_NO_MODIFICATION_ALLOWED: () => [number, number];
+    readonly ERR_QUOTA_EXCEEDED: () => [number, number];
+    readonly ERR_TYPE_MISMATCH: () => [number, number];
     readonly __wbg_direntry_free: (a: number, b: number) => void;
     readonly __wbg_fs_free: (a: number, b: number) => void;
     readonly __wbg_get_direntry_name: (a: number) => [number, number];
@@ -254,28 +273,6 @@ export interface InitOutput {
     readonly project_sigMd5: (a: any) => any;
     readonly project_updateInfoSubscribe: (a: number, b: any) => void;
     readonly project_writeAllToDisk: (a: any) => void;
-    readonly getWasmMemory: () => any;
-    readonly getWasmModule: () => any;
-    readonly ERR_ABORT: () => [number, number];
-    readonly ERR_INVALID_STATE: () => [number, number];
-    readonly ERR_KEY_ALREADY_EXISTS: () => [number, number];
-    readonly ERR_NOT_ALLOWED: () => [number, number];
-    readonly ERR_NOT_FOUND: () => [number, number];
-    readonly ERR_NO_MODIFICATION_ALLOWED: () => [number, number];
-    readonly ERR_QUOTA_EXCEEDED: () => [number, number];
-    readonly ERR_TYPE_MISMATCH: () => [number, number];
-    readonly __wbg_buildoptions_free: (a: number, b: number) => void;
-    readonly __wbg_get_buildoptions_cleanup: (a: number) => number;
-    readonly __wbg_roottask_free: (a: number, b: number) => void;
-    readonly __wbg_set_buildoptions_cleanup: (a: number, b: number) => void;
-    readonly buildoptions_config: (a: number) => any;
-    readonly buildoptions_new: () => number;
-    readonly buildoptions_set_config: (a: number, b: any) => void;
-    readonly registerWorkerScheduler: (a: any, b: any) => void;
-    readonly workerCreated: (a: number) => void;
-    readonly rust_mi_get_default_heap: () => number;
-    readonly rust_mi_get_thread_id: () => number;
-    readonly rust_mi_set_default_heap: (a: number) => void;
     readonly __wbg_get_wasmtaskmessage_taskId: (a: number) => number;
     readonly __wbg_get_webworkercreation_options: (a: number) => number;
     readonly __wbg_get_webworkeroptions_cwd: (a: number) => [number, number];
